ref(integrations): rename module task
