+++ src/sentry/tasks/integrations/__init__.py
@@ -38,7 +38,7 @@ __all__ = (
 settings.CELERY_IMPORTS += (
     "sentry.tasks.integrations.create_comment",
     "sentry.tasks.integrations.github.pr_comment",
-    "sentry.tasks.integrations.kick_off_status_syncs",
+    "sentry.tasks.integrations.kick_off_status_syncs_impl",
     "sentry.tasks.integrations.link_all_repos",
     "sentry.tasks.integrations.migrate_opsgenie_plugins",
     "sentry.tasks.integrations.migrate_issues",
@@ -53,7 +53,7 @@ settings.CELERY_IMPORTS += (
 )
 
 from .create_comment import create_comment
-from .kick_off_status_syncs import kick_off_status_syncs
+from .kick_off_status_syncs_impl import kick_off_status_syncs
 from .migrate_issues import migrate_issues
 from .migrate_opsgenie_plugins import migrate_opsgenie_plugin
 from .migrate_repo import migrate_repo
